Release v1.0.0
- documentation marks v0.1-v1.0 as released and states the stable schema
- lint cleanups in large/duplicates/volumes/args
- version 1.0.0

+++ src/core/volumes.rs
@@ -44,9 +44,9 @@ pub fn volumes() -> Result<Vec<VolumeInfo>> {
             continue;
         }
         let block = fs.f_bsize as u64;
-        let total = fs.f_blocks as u64 * block;
-        let free = fs.f_bfree as u64 * block;
-        let available = fs.f_bavail as u64 * block;
+        let total = fs.f_blocks * block;
+        let free = fs.f_bfree * block;
+        let available = fs.f_bavail * block;
         let is_system =
             mount_point.starts_with("/System/") || mount_point == "/System/Volumes/Data";
         out.push(VolumeInfo {
